bo_code/requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
future==0.18.1
kiwisolver==1.1.0
matplotlib==2.2.4
numpy==1.16.5
pudb==2019.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
scipy==1.2.2
six==1.12.0
subprocess32==3.5.4
tqdm==4.36.1
urwid==2.0.1
dragonfly-opt==0.1.4
